Refactor Prettier config to support Vue-specific settings
Split Prettier configuration into base and Vue-specific configs. Updated package.json to use the new Vue config in the Vue package. Adjusted exports in tooling/prettier/package.json to expose both base and Vue configs, and removed the old index.mjs file.
