+++ packages/vue/package.json
@@ -43,5 +43,5 @@
     "vue": "^3.4.0",
     "vue-tsc": "^1.8.0"
   },
-  "prettier": "@card-stack/prettier-config"
+  "prettier": "@card-stack/prettier-config/vue"
 }
