+++ tooling/prettier/base.mjs
@@ -1,0 +1,13 @@
+/** @type {import("prettier").Config} */
+const config = {
+  semi: false,
+  singleQuote: true,
+  tabWidth: 2,
+  trailingComma: 'none',
+  printWidth: 100,
+  bracketSpacing: true,
+  endOfLine: 'lf',
+  plugins: ['@trivago/prettier-plugin-sort-imports', 'prettier-plugin-tailwindcss']
+}
+
+export default config
